moved insertUser to AuthController
